server/end/requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
Pillow==10.2.0
torch==2.2.1
torchvision==0.17.1 
numpy==1.26.1